Allow loading of vendor-<class> hardware modules.
If the requested class for a module to be loaded is prefixed with
"vendor-", check that the module class matches the portion after
"vendor-" instead of comparing the full name of the requested class
to the module class. This allows easy use of the libhardware loader
in modules that wrap vendor binaries, which can directly use
hw_get_module to load the vendor binary.

Change-Id: I09f5594589e27c62448ef49ee47225b57dc6a714